Rounded seconds in user feedback

+++ Main.cpp
@@ -61,7 +61,7 @@ void ShowProgress(float num, float denom) {
 
 void PrintExecutionTime(float time) {
   int min = int(time / CLOCKS_PER_SEC) / 60;
-  float sec = (time / CLOCKS_PER_SEC) - (min * 60);
+  int sec = int(time / CLOCKS_PER_SEC) - (min * 60);
   cout << "\n------------ TIME ---------------" << endl;
   cout << "Rendered file for " << min << "min " << sec << "s" << endl;
   cout << "---------------------------------" << endl;
